feat: add QuoteList component for managing quotes in the dashboard and quotes page
- Implemented QuoteList component to display, create, edit, view, and delete quotes.
- Integrated QuoteList into the dashboard and quotes pages, replacing existing inline logic.
- Added pagination support for quotes and invoices.
- Updated dashboard to show latest quotes and invoices using the new components.
- Refactored invoice management to use InvoiceList component for consistency.
- Improved loading and empty state handling in both quotes and invoices sections.
- Removed unused dialog components and API calls related to quotes and invoices.
