+++ frontend/src/pages/(app)/quotes/index.tsx
@@ -1,31 +1,20 @@
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { Download, Edit, Eye, FileSignature, FileText, Mail, Phone, Plus, Search, Signature, Trash2 } from "lucide-react"
-import { useEffect, useState } from "react"
-import { useGet, useGetRaw, usePost } from "@/lib/utils"
+import { Card, CardContent } from "@/components/ui/card"
+import { FileSignature, Plus, Search } from "lucide-react"
+import { useEffect, useRef, useState } from "react"
+import { useGet, useGetRaw } from "@/lib/utils"
 
-import BetterPagination from "@/components/pagination"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import type { Quote } from "@/types"
-import { QuoteCreate } from "./_components/quote-create"
-import { QuoteDeleteDialog } from "./_components/quote-delete"
-import { QuoteEdit } from "./_components/quote-edit"
-import { QuotePdfModal } from "./_components/quote-pdf-view"
-import { QuoteViewDialog } from "./_components/quote-view"
+import { QuoteList } from "@/components/quote-list"
+import type { QuoteListHandle } from "@/components/quote-list"
 
 export default function Quotes() {
-    const { trigger: triggerMarkAsSigned } = usePost(`/api/quotes/mark-as-signed`)
-    const { trigger: triggerCreateInvoice } = usePost(`/api/invoices/create-from-quote`)
-
+    const quoteListRef = useRef<QuoteListHandle>(null)
     const [page, setPage] = useState(1)
 
     const { data: quotes, mutate, loading } = useGet<{ pageCount: number, quotes: Quote[] }>(`/api/quotes?page=${page}`)
 
-    const [createQuoteDialog, setCreateQuoteDialog] = useState<boolean>(false)
-    const [editQuoteDialog, setEditQuoteDialog] = useState<Quote | null>(null)
-    const [viewQuoteDialog, setViewQuoteDialog] = useState<Quote | null>(null)
-    const [viewQuotePdfDialog, setViewQuotePdfDialog] = useState<Quote | null>(null)
-    const [deleteQuoteDialog, setDeleteQuoteDialog] = useState<Quote | null>(null)
     const [downloadQuotePdf, setDownloadQuotePdf] = useState<Quote | null>(null)
 
     const { data: pdf } = useGetRaw<Response>(`/api/quotes/${downloadQuotePdf?.id}/pdf`)
@@ -54,41 +43,28 @@ export default function Quotes() {
         quote.status.toLowerCase().includes(searchTerm.toLowerCase())
     ) || []
 
-    function handleAddClick() {
-        setCreateQuoteDialog(true)
-    }
-
-    function handleEdit(quote: Quote) {
-        setEditQuoteDialog(quote)
-    }
-
-    function handleView(quote: Quote) {
-        setViewQuoteDialog(quote)
-    }
-
-    function handleViewPdf(quote: Quote) {
-        setViewQuotePdfDialog(quote)
-    }
-
-    function handleDownloadPdf(quote: Quote) {
-        setDownloadQuotePdf(quote)
-    }
-
-    function handleDelete(quote: Quote) {
-        setDeleteQuoteDialog(quote)
-    }
-
-    function handleMarkAsSigned(quoteId: string) {
-        triggerMarkAsSigned({ id: quoteId }).then(() => {
-            mutate();
-        }).catch((error) => {
-            console.error("Error marking quote as signed:", error);
-        });
-    }
-
-    function handleCreateInvoice(quoteId: string) {
-        triggerCreateInvoice({ quoteId })
-    }
+    const emptyState = (
+        <div className="text-center py-12">
+            <FileSignature className="mx-auto h-12 w-12 text-gray-400" />
+            <h3 className="mt-2 text-sm font-medium text-foreground">
+                {searchTerm ? 'No quotes found' : 'No quotes added yet'}
+            </h3>
+            <p className="mt-1 text-sm text-primary">
+                {searchTerm
+                    ? 'Try a different search term'
+                    : 'Start adding quotes to manage your business effectively.'
+                }
+            </p>
+            {!searchTerm && (
+                <div className="mt-6">
+                    <Button onClick={() => quoteListRef.current?.handleAddClick()}>
+                        <Plus className="h-4 w-4 mr-2" />
+                        Add New Quote
+                    </Button>
+                </div>
+            )}
+        </div>
+    )
 
     return (
         <div className="max-w-6xl mx-auto space-y-6 p-6">
@@ -118,7 +94,7 @@ export default function Quotes() {
                     </div>
 
                     <Button
-                        onClick={handleAddClick}
+                        onClick={() => quoteListRef.current?.handleAddClick()}
                     >
                         <Plus className="h-4 w-4 mr-0 md:mr-2" />
                         <span className="hidden md:inline-flex">
@@ -180,186 +156,18 @@ export default function Quotes() {
                 </Card>
             </div>
 
-            <Card className="gap-0">
-                <CardHeader className="border-b">
-                    <CardTitle className="flex items-center space-x-2">
-                        <FileSignature className="h-5 w-5 " />
-                        <span>Quotes</span>
-                    </CardTitle>
-                    <CardDescription>Manage your quotes, view details, edit or delete them.</CardDescription>
-                </CardHeader>
-                <CardContent className="p-0">
-                    {loading && (
-                        <div className="flex items-center justify-center py-12">
-                            <div className="animate-spin rounded-full h-16 w-16 border-b-2 border-blue-500"></div>
-                        </div>
-                    )}
-                    {!loading && filteredQuotes.length === 0 ? (
-                        <div className="text-center py-12">
-                            <FileSignature className="mx-auto h-12 w-12 text-gray-400" />
-                            <h3 className="mt-2 text-sm font-medium text-foreground">
-                                {searchTerm ? 'No quotes found' : 'No quotes added yet'}
-                            </h3>
-                            <p className="mt-1 text-sm text-primary">
-                                {searchTerm
-                                    ? 'Try a different search term'
-                                    : 'Start adding quotes to manage your business effectively.'
-                                }
-                            </p>
-                            {!searchTerm && (
-                                <div className="mt-6">
-                                    <Button onClick={handleAddClick}>
-                                        <Plus className="h-4 w-4 mr-2" />
-                                        Add New Quote
-                                    </Button>
-                                </div>
-                            )}
-                        </div>
-                    ) : (
-                        <div className="divide-y">
-                            {filteredQuotes.map((quote, index) => (
-                                <div key={index} className="p-4 sm:p-6">
-                                    <div className="flex flex-row sm:items-center sm:justify-between gap-4">
-                                        <div className="flex flex-row items-center gap-4 w-full">
-                                            <div className="p-2 bg-blue-100 rounded-lg mb-4 md:mb-0 w-fit h-fit">
-                                                <FileSignature className="h-5 w-5 text-blue-600" />
-                                            </div>
-                                            <div className="flex-1">
-                                                <div className="flex flex-wrap items-center gap-2">
-                                                    <h3 className="font-medium text-foreground break-words">
-                                                        {quote.title || `Quote #${quote.number}`}
-                                                    </h3>
-                                                    <span className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-semibold
-                ${quote.status === 'DRAFT' ? 'bg-yellow-100 text-yellow-800' :
-                                                            quote.status === 'SIGNED' ? 'bg-blue-100 text-blue-800' :
-                                                                quote.status === 'EXPIRED' ? 'bg-red-100 text-red-800' :
-                                                                    quote.status === 'SENT' ? 'bg-green-100 text-green-800' :
-                                                                        'bg-gray-100 text-gray-800'
-                                                        }`}>
-                                                        {quote.status}
-                                                    </span>
-                                                </div>
-                                                <div className="mt-2 flex flex-col sm:flex-row flex-wrap gap-2 text-sm text-primary">
-                                                    <div className="flex items-center space-x-1">
-                                                        <Mail className="h-4 w-4" />
-                                                        <span className="break-all">{quote.client.contactEmail}</span>
-                                                    </div>
-                                                    {quote.client.contactPhone && (
-                                                        <div className="flex items-center space-x-1">
-                                                            <Phone className="h-4 w-4" />
-                                                            <span>{quote.client.contactPhone}</span>
-                                                        </div>
-                                                    )}
-                                                </div>
-                                            </div>
-                                        </div>
-
-                                        <div className="grid grid-cols-2 lg:flex justify-start sm:justify-end gap-2">
-                                            <Button
-                                                tooltip="View Quote"
-                                                variant="ghost"
-                                                size="icon"
-                                                onClick={() => handleView(quote)}
-                                                className="text-gray-600 hover:text-blue-600"
-                                            >
-                                                <Eye className="h-4 w-4" />
-                                            </Button>
-                                            <Button
-                                                tooltip="View PDF"
-                                                variant="ghost"
-                                                size="icon"
-                                                onClick={() => handleViewPdf(quote)}
-                                                className="text-gray-600 hover:text-pink-600"
-                                            >
-                                                <FileText className="h-4 w-4" />
-                                            </Button>
-                                            <Button
-                                                tooltip="Download PDF"
-                                                variant="ghost"
-                                                size="icon"
-                                                onClick={() => handleDownloadPdf(quote)}
-                                                className="text-gray-600 hover:text-amber-600"
-                                            >
-                                                <Download className="h-4 w-4" />
-                                            </Button>
-                                            <Button
-                                                tooltip="Edit Quote"
-                                                variant="ghost"
-                                                size="icon"
-                                                onClick={() => handleEdit(quote)}
-                                                className="text-gray-600 hover:text-green-600"
-                                            >
-                                                <Edit className="h-4 w-4" />
-                                            </Button>
-                                            {quote.status !== 'SIGNED' && (
-                                                <Button
-                                                    tooltip="Mark as Signed"
-                                                    variant="ghost"
-                                                    size="icon"
-                                                    onClick={() => handleMarkAsSigned(quote.id)}
-                                                    className="text-gray-600 hover:text-blue-600"
-                                                >
-                                                    <Signature className="h-4 w-4" />
-                                                </Button>
-                                            )}
-                                            {quote.status === 'SIGNED' && (
-                                                <Button
-                                                    tooltip="Create Invoice"
-                                                    variant="ghost"
-                                                    size="icon"
-                                                    onClick={() => handleCreateInvoice(quote.id)}
-                                                    className="text-gray-600 hover:text-green-600"
-                                                >
-                                                    <FileText className="h-4 w-4" />
-                                                </Button>
-                                            )}
-                                            <Button
-                                                tooltip="Delete Quote"
-                                                variant="ghost"
-                                                size="icon"
-                                                onClick={() => handleDelete(quote)}
-                                                className="text-gray-600 hover:text-red-600"
-                                            >
-                                                <Trash2 className="h-4 w-4" />
-                                            </Button>
-                                        </div>
-                                    </div>
-                                </div>
-                            ))}
-                        </div>
-
-                    )}
-                </CardContent>
-                <CardFooter>
-                    {!loading && filteredQuotes.length > 0 && (
-                        <BetterPagination pageCount={quotes?.pageCount || 1} page={page} setPage={setPage} />
-                    )}
-                </CardFooter>
-            </Card>
-
-            <QuoteCreate
-                open={createQuoteDialog}
-                onOpenChange={(open) => { setCreateQuoteDialog(open); mutate() }}
-            />
-
-            <QuoteEdit
-                quote={editQuoteDialog}
-                onOpenChange={(open) => { if (!open) setEditQuoteDialog(null); mutate() }}
-            />
-
-            <QuoteViewDialog
-                quote={viewQuoteDialog}
-                onOpenChange={(open) => { if (!open) setViewQuoteDialog(null) }}
-            />
-
-            <QuotePdfModal
-                quote={viewQuotePdfDialog}
-                onOpenChange={(open) => { if (!open) setViewQuotePdfDialog(null) }}
-            />
-
-            <QuoteDeleteDialog
-                quote={deleteQuoteDialog}
-                onOpenChange={(open) => { if (!open) setDeleteQuoteDialog(null); mutate() }}
+            <QuoteList
+                ref={quoteListRef}
+                quotes={filteredQuotes}
+                loading={loading}
+                title="Quotes"
+                description="Manage your quotes, view details, edit or delete them."
+                page={page}
+                pageCount={quotes?.pageCount || 1}
+                setPage={setPage}
+                mutate={mutate}
+                emptyState={emptyState}
+                showCreateButton={true}
             />
 
         </div>
